bugfix: always update weights for current day only.
Scenario 1:
The bug would happen if on day 1 user would not have any data.
On day 2, script will try to fetch data for both days 1 and 2.
If user had less weight on day 1 afternoon, this would be written for day 2.
The measurement from day 1 afternoon is **not** discarded and is written for
day 2. This is unexpected behavior.

Scenario 2:
User has data on day 1.
User makes a measurement on day 1 afternoon.
On day 2, we fetch data for day 2 only.
The measurement from day 1 afternoon is discarded as expected.

--

The mechanism that fetches everything from last successful fetch is a leftover
from previous implementation where we would be able to change measurements from
previous dates. This turned out to be confusing when people buy/sell weights
and something in the past changes.